Update the propagation rule of a function call and pass the ifelse test case successfully

+++ tests/bound/ifelse/naive_bound/naive/ifelse.c
@@ -2,7 +2,7 @@
 int64_t _f_(int64_t x){
 	int64_t r = 0;
 	uint16_t _2 = 0;
-	uint16_t _3 = 0;
+	int64_t _3 = 0;
 	uint16_t _4 = 0;
 	int64_t _5 = 0;
 	uint16_t _6 = 0;
